+++ packages/bridge-controller/src/index.ts
@@ -143,6 +143,7 @@ export {
   isNativeAddress,
   isSolanaChainId,
   isBitcoinChainId,
+  isStellarChainId,
   isTronChainId,
   isNonEvmChainId,
   getNativeAssetForChainId,